Added custom hooks
- Added useQuery hook
- Added useMutation hook

Note: These hooks were added because with apollo-server's formatResponse and formatError, the client doesn't ply well. It was stuck in an infinite loading initially, and now it returns data and error as undefined.

To overcome this, these two custom hooks were added. These hooks play identically to the original hooks, except if there is only a singular data or a singular error returned, it pops the array and makes data accessible as data.<field> instead of data.<query/mutation name>.<field>. Similarly with errors.